requirements.txt
asgiref==3.8.1
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.41.0
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==5.1
django-allauth==0.61.1
django-cloudinary-storage==0.3.0
django-filter==24.3
django-summernote==0.8.20.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.53.0
gunicorn==22.0.0
idna==3.7
kiwisolver==1.4.5
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.2
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.7.0
